+++ src/main/java/com/bobmowzie/mowziesmobs/server/item/ItemWroughtHelm.java
@@ -117,7 +117,7 @@ public class ItemWroughtHelm extends ArmorItem {
 
         @Override
         public float getToughness() {
-            return ConfigHandler.COMMON.TOOLS_AND_ABILITIES.WROUGHT_HELM.armorConfig.toughness.get();
+            return ConfigHandler.COMMON.TOOLS_AND_ABILITIES.WROUGHT_HELM.armorConfig.toughness.get().floatValue();
         }
 
         @Override
